✨ feat: updated sent ccip token v2
