+++ addressList/mumbai.json
@@ -1,1 +1,1 @@
-{"AstroDogNft":"0x84d1242291dA9bd26613B86003aB48a696F5AB05","mumbai":"0xCAcCF9A29511Dc57296A284fd1894F396aFc1a58"}
+{"AstroDogNft":"0x84d1242291dA9bd26613B86003aB48a696F5AB05","mumbai":"0x1739fB64bC4d4Cad7AC2b5D20dE30C374C1dec5E"}
