+++ addressList/sepolia.json
@@ -1,1 +1,1 @@
-{"sepolia":"0x1C9CdFB6715110ba65341B16004dA1f7aa373Ac7","BadApeNft":"0x16bC29a24f74FB915f78eB7d2104684CaD3356b6"}
+{"sepolia":"0x633BfAe6736515b945eA701dD76D7D7e324404fe","BadApeNft":"0x16bC29a24f74FB915f78eB7d2104684CaD3356b6"}
