+++ scripts/tradeLinkCCIPV1.deploy.ts
